Fix issues with optional capturing groups
In regular expressions like a(x)?b the $1 capturing group won't be present
if x wasn't matched.

To handle this I now changes the result format for capturing groups. They
will be put into the fourth element of the token. If no capturing groups are
present or none matched the fourth element won't be added.

The matches in the forth element do not contain the 0 offset, because it is
already included as the third element.